+++ packages/core/src/components/fields/LayoutMultiSchemaField.tsx
@@ -104,7 +104,6 @@ export default function LayoutMultiSchemaField<
     registry,
     uiSchema,
     schema,
-    formContext,
     autofocus,
     readonly,
     required,
@@ -206,7 +205,6 @@ export default function LayoutMultiSchemaField<
         label={(title || schema.title) ?? ''}
         disabled={disabled || (Array.isArray(enumOptions) && isEmpty(enumOptions))}
         uiSchema={uiSchema}
-        formContext={formContext}
         autofocus={autofocus}
         readonly={readonly}
         required={required}